refactor many things
