--- packages/nextjs/components/scaffold-eth/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Input";
