+++ packages/nextjs/components/scaffold-move/Input/Bytes32Input.tsx
@@ -1,6 +1,6 @@
 import { useCallback } from "react";
 import { hexToString, isHex, stringToHex } from "viem";
-import { CommonInputProps, InputBase } from "~~/components/scaffold-eth";
+import { CommonInputProps, InputBase } from "~~/components/scaffold-move";
 
 export const Bytes32Input = ({ value, onChange, name, placeholder, disabled }: CommonInputProps) => {
   const convertStringToBytes32 = useCallback(() => {
