+++ packages/nextjs/components/scaffold-move/index.tsx
@@ -2,3 +2,4 @@ export * from "./Address";
 export * from "./Balance";
 export * from "./CustomConnectButton";
 export * from "./BlockieAvatar";
+export * from "./Input";
